refactor: Move `display_list` under `renderer`
